+++ mkosi/distributions/custom.py
@@ -18,6 +18,10 @@ class Installer(DistributionInstaller):
         pass
 
     @classmethod
+    def sync(cls, context: Context) -> None:
+        pass
+
+    @classmethod
     def install(cls, context: Context) -> None:
         pass