+++ mkosi/distributions/mageia.py
@@ -6,7 +6,6 @@ from collections.abc import Iterable, Sequence
 from mkosi.config import Architecture
 from mkosi.context import Context
 from mkosi.distributions import Distribution, fedora, join_mirror
-from mkosi.installer.dnf import Dnf
 from mkosi.installer.rpm import RpmRepository, find_rpm_gpgkey
 from mkosi.log import die
 
@@ -50,9 +49,6 @@ class Installer(fedora.Installer):
         if context.config.local_mirror:
             yield RpmRepository("core-release", f"baseurl={context.config.local_mirror}", gpgurls)
             return
-
-        if context.want_local_repo():
-            yield Dnf.localrepo()
 
         if context.config.mirror:
             url = f"baseurl={join_mirror(context.config.mirror, 'distrib/$releasever/$basearch/media/core/')}"