+++ mkosi/distributions/openmandriva.py
@@ -6,7 +6,6 @@ from collections.abc import Iterable, Sequence
 from mkosi.config import Architecture
 from mkosi.context import Context
 from mkosi.distributions import Distribution, fedora, join_mirror
-from mkosi.installer.dnf import Dnf
 from mkosi.installer.rpm import RpmRepository, find_rpm_gpgkey
 from mkosi.log import die
 
@@ -56,9 +55,6 @@ class Installer(fedora.Installer):
         if context.config.local_mirror:
             yield RpmRepository("main-release", f"baseurl={context.config.local_mirror}", gpgurls)
             return
-
-        if context.want_local_repo():
-            yield Dnf.localrepo()
 
         url = f"baseurl={join_mirror(mirror, '$releasever/repository/$basearch/main')}"
         yield RpmRepository("main-release", f"{url}/release", gpgurls)