+++ mkosi/installer/__init__.py
@@ -1,20 +1,60 @@
 # SPDX-License-Identifier: LGPL-2.1+
 
+import os
 from pathlib import Path
 
-from mkosi.config import ConfigFeature
+from mkosi.config import Config, ConfigFeature, OutputFormat
 from mkosi.context import Context
 from mkosi.run import find_binary
 from mkosi.sandbox import finalize_crypto_mounts
-from mkosi.tree import rmtree
+from mkosi.tree import move_tree, rmtree
 from mkosi.types import PathString
 from mkosi.util import flatten
 
 
 class PackageManager:
     @classmethod
+    def subdir(cls, config: Config) -> Path:
+        raise NotImplementedError
+
+    @classmethod
+    def cache_subdirs(cls, cache: Path) -> list[Path]:
+        raise NotImplementedError
+
+    @classmethod
     def scripts(cls, context: Context) -> dict[str, list[PathString]]:
         raise NotImplementedError
+
+    @classmethod
+    def mounts(cls, context: Context) -> list[PathString]:
+        mounts: list[PathString] = [
+            *(["--ro-bind", m, m] if (m := context.config.local_mirror) else []),
+            *finalize_crypto_mounts(tools=context.config.tools()),
+            "--bind", context.packages, "/work/packages",
+        ]
+
+        subdir = context.config.distribution.package_manager(context.config).subdir(context.config)
+
+        for d in ("cache", "lib"):
+            src = context.package_cache_dir / d / subdir
+            mounts += ["--bind", src, Path("/var") / d / subdir]
+
+            # If we're not operating on the configured package cache directory, we're operating on a snapshot of the
+            # repository metadata in the image root directory. To make sure any downloaded packages are still cached in
+            # the configured package cache directory in this scenario, we mount in the relevant directories from the
+            # configured package cache directory.
+            if d == "cache" and context.package_cache_dir != context.config.package_cache_dir_or_default():
+                caches = context.config.distribution.package_manager(context.config).cache_subdirs(src)
+                mounts += flatten(
+                    [
+                        "--bind",
+                        os.fspath(context.config.package_cache_dir_or_default() / d / subdir / p.relative_to(src)),
+                        Path("/var") / d / subdir / p.relative_to(src),
+                    ]
+                    for p in caches
+                )
+
+        return mounts
 
 
 def clean_package_manager_metadata(context: Context) -> None:
@@ -24,6 +64,25 @@ def clean_package_manager_metadata(context: Context) -> None:
     Try them all regardless of the distro: metadata is only removed if
     the package manager is not present in the image.
     """
+    if (
+        context.package_cache_dir.is_relative_to(context.root) and
+        not context.config.overlay and (
+            context.config.clean_package_metadata != ConfigFeature.disabled or
+            context.config.output_format not in (OutputFormat.directory, OutputFormat.tar)
+        )
+    ):
+        # Instead of removing the package cache directory from the image, we move it to the workspace so it stays
+        # available for later steps and is automatically removed along with the workspace when the build finishes.
+        context.package_cache_dir = move_tree(
+            context.package_cache_dir, context.workspace / "package-cache-dir",
+            tools=context.config.tools(),
+            sandbox=context.sandbox(
+                options=[
+                    "--bind", context.package_cache_dir.parent, context.package_cache_dir.parent,
+                    "--bind", context.workspace, context.workspace,
+                ],
+            ),
+        )
 
     if context.config.clean_package_metadata == ConfigFeature.disabled:
         return
@@ -37,28 +96,3 @@ def clean_package_manager_metadata(context: Context) -> None:
         if always or not find_binary(tool, root=context.root):
             rmtree(*(context.root / p for p in paths),
                    sandbox=context.sandbox(options=["--bind", context.root, context.root]))
-
-
-def finalize_package_manager_mounts(context: Context) -> list[PathString]:
-    from mkosi.installer.dnf import Dnf
-
-    mounts: list[PathString] = [
-        *(["--ro-bind", m, m] if (m := context.config.local_mirror) else []),
-        *finalize_crypto_mounts(tools=context.config.tools()),
-        "--bind", context.packages, "/work/packages",
-    ]
-
-    mounts += flatten(
-        ["--bind", context.cache_dir / d, Path("/var") / d]
-        for d in (
-            "lib/apt",
-            "cache/apt",
-            f"cache/{Dnf.subdir(context.config)}",
-            f"lib/{Dnf.subdir(context.config)}",
-            "cache/pacman/pkg",
-            "cache/zypp",
-        )
-        if (context.cache_dir / d).exists()
-    )
-
-    return mounts